Added Recycler Behavior/ Player Controller Script
Once again, very primitive scripts. What (should) happen is the Recycler enemy begins in its idle gnawing animation, and once the player is in proximity to it, it will use the walking forward animation to go towards the player, and once the Recycler is close to the player it will attack. Still need to figure out damage logic. Also added a Player Controller script that basically does nothing :)